Add files via upload
Shown is the calibration data and manipulating it to get sensor fusion. As you can see by the first 3 graphs the data gets more and more compressed until all are at approximately the same range of values.

The last graph demonstrates the weights of the data. If the value of the IR sensor is too high or too low, the car will turn more drastically, if the sensor value is more mellow a sharper correction to stay on the path is not needed. 

As you can see in the code some of these final values were used to successfully implement proportional integral derivative control.